pyproject.toml
[build-system]
requires = [
    "setuptools>=61.2",
    "setuptools_scm[toml]>=7"
]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true
zip-safe = false
license-files = [
    "LICENSE.md",
]
py-modules = []

[tool.setuptools.package-dir]
"" = "workflow/src"

[tool.setuptools.packages.find]
where = ["workflow/src"]

[tool.setuptools_scm]
write_to = "workflow/src/legendsimflow/_version.py"

[project]
name = "legend_simflow"
description = "Python package for processing LEGEND simulations"
authors = [
    {name = "Luigi Pertoldi", email = "gipert@pm.me"},
]
maintainers = [
    {name = "The LEGEND Collaboration"},
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX",
    "Operating System :: Unix",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Scientific/Engineering",
]
readme = "README.md"
requires-python = ">=3.11"
dynamic = ["version"]

dependencies = [
    "colorlog",
    "pylegendmeta",
    "pip",
    "legend-dataflow-scripts >=0.1.2",
    "pyyaml",
]

[project.optional-dependencies]
# these are needed to run the data production
runprod = [
    "pygments",
    "snakemake>=8.16",
]
test = [
    "legend_simflow[runprod]",
    "pytest>=6",
    "pytest-cov>=3",
]
dev = [
    "legend_simflow[runprod,test]",
    "pre-commit",
]
docs = [
    "sphinx>=7.0",
    "myst_parser>=0.13",
    "sphinx_inline_tabs",
    "sphinx_copybutton",
    "sphinx_autodoc_typehints",
    "furo>=2023.08.17",
]

[tool.uv.workspace]
exclude = ["generated", "inputs", "software", "workflow"]

[tool.uv]
default-groups = []

[tool.pytest.ini_options]
minversion = "6.0"
addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
xfail_strict = true
filterwarnings = [
  "error",
]
log_cli_level = "INFO"
testpaths = [
  "tests",
]

[tool.ruff]
src = ["workflow/src"]

[tool.ruff.lint]
extend-select = [
  "ARG",      # flake8-unused-arguments
  "B",        # flake8-bugbear
  "C4",       # flake8-comprehensions
  "EM",       # flake8-errmsg
  "EXE",      # flake8-executable
  "G",        # flake8-logging-format
  "I",        # isort
  "ICN",      # flake8-import-conventions
  "NPY",      # NumPy specific rules
  "PD",       # pandas-vet
  "PGH",      # pygrep-hooks
  "PIE",      # flake8-pie
  "PL",       # pylint
  "PT",       # flake8-pytest-style
  "PTH",      # flake8-use-pathlib
  "RET",      # flake8-return
  "RUF",      # Ruff-specific
  "SIM",      # flake8-simplify
  "T20",      # flake8-print
  "UP",       # pyupgrade
  "YTT",      # flake8-2020
]
ignore = [
  "PT011",    # `pytest.raises(ValueError)` is too broad
  "PLR09",    # Too many <...>
  "PLR2004",  # Magic value used in comparison
  "ISC001",   # Conflicts with formatter
]
isort.required-imports = ["from __future__ import annotations"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["T20"]
"noxfile.py" = ["T20"]


[tool.pylint]
py-version = "3.9"
ignore-paths = [".*/_version.py"]
reports.output-format = "colorized"
similarities.ignore-imports = "yes"
messages_control.disable = [
  "design",
  "fixme",
  "line-too-long",
  "missing-module-docstring",
  "missing-function-docstring",
  "wrong-import-position",
  "too-many-nested-blocks"
]
